pyproject.toml
[project]
name = "azure-content-safety"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-ai-contentsafety>=1.0.0",
    "azure-core>=1.37.0",
    "azure-identity>=1.25.1",
    "lagom>=2.7.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.0",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config", "-ra"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "azure_content_safety/hosting.py",
    "azure_content_safety/protocols/*.py", 
]
